Add --twisted to bpython.urwid and use it for --plugin and --server.
This uses the currently installed reactor (and installs the default
implicitly if necessary), as opposed to explicitly installing a
default reactor. This is useful for some bpython embedding usecases
where a reactor is already installed by the time bpython.urwid.main()
is called, and it fixes a bug where bpython defaults to the select
reactor instead of twisted's default.