Search for import completion modules in idle generator
A different technique that amygdalama and I talked about, but required
modifications to curtsies: an idle function like cli.py uses.

I pulled from amygdalama's repo because I thought we would use that code, and it
will be useful when we add Windows support, but it's not necessary right now.