fix #432 bpython-curtsies crashes if no readline history on enter
It actually works correctly if there is no existing history the first time you press enter.
But then insert_into_history sets entries to [] and tries to read from the history file,
which will do nothing if the history file is empty.

History should be fixed properly on master, but for now just fixing the
call which shouldn't have been rl_history.last() in the first place.