fix bpython-curtsies crashes if no readline history on enter #432
It actually works correctly if there is no existing history the first time you press enter. But then insert_into_history sets entries to [] and tries to read from the history file, which will do nothing if the history file is empty.

This is fixed properly on master, for now just change last() to return an empty string if history entries are empty.