Fix inspection of built-in functions with >= 3.11
The built-in functions no longer have their signature in the docstring,
but now inspect.signature can produce results. But as we have no source
for built-in functions, we cannot replace the default values. Hence, we
handle built-in functions in an extra step.

This commit also changes the handling of default values slightly. They
are now always put into a _Repr.