More general adaptation of showsyntaxerror() to Python 3.13
Python 3.13's code.InteractiveInterpreter adds a new **kwargs
argument to its showsyntaxerror() method. Currently, the only
use of it is to send a named argument of name "source". Whilst
the current adapation of repl.Interpreter is specific and should
work in the short term, here is a more general solution.